+++ CHANGELOG.md
@@ -14,6 +14,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [1.2.0]
 
 ### Added
+- `acli` | `1.3.0` | Added `~/.acli.yaml` file managed by viper for persistent
+  configuration.
 
 ### Changed